+++ client/src/components/RandomAudioActvity.js
@@ -1,19 +1,12 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
-import { get, reduce } from 'lodash';
+import { get, reduce, each } from 'lodash';
 import { Button } from 'react-bootstrap';
 import { FormControl, ControlLabel } from 'react-bootstrap';
 import Sound from 'react-sound';
 
 import { Header, PlainText } from '../components/Typography';
 import { idFromString } from '../util';
-
-const FormLayout = styled.div`
-  margin: auto;
-  width: 900px;
-  padding: 0px 15px;
-  height: 100%;
-`;
 
 const IntroText = styled(PlainText)`
   margin-left: 0px;
@@ -27,31 +20,33 @@ const StyledFormControl = styled(FormControl)`
   margin-bottom: 15px;
 `;
 
+const setUpQuestions = (audioFiles, questions) => {
+  const toReturn = {};
+  each(questions, (question) => {
+    each(audioFiles, (file) => {
+      const questionId = `${file.split('.')[0]}.${idFromString(question.question)}`;
+      const questionData = {
+        type: question.type,
+        audioFile: file
+      };
+      toReturn[questionId] = questionData;
+    });
+  });
+  return toReturn;
+};
+
 class RandomAudioActivity extends Component {
 
-  state ={
-    audioState: Sound.status.PAUSED
+  state = {
+    audioState: Sound.status.PAUSED,
+    questions: setUpQuestions(
+      this.props.activity.audioFiles,
+      this.props.activity.questions
+    ),
+    // Activities start on the first question, incomplete
+    questionIndex: 0,
+    activityComplete: false
   }
-
-  // state = {
-  //   questions: reduce(this.props.activity.questions, (result, value, key) => {
-  //     const questionData = {
-  //       type: value.type,
-  //       indexInActivity: key
-  //     };
-  //     const questionId = idFromString(value.question);
-  //     if (value.type === 'text'){
-  //       questionData['response'] = '';
-  //     }
-  //     if (value.type === 'select'){
-  //       questionData['response'] = get(value, 'options[0]', '');
-  //     }
-  //     return {
-  //       ...result,
-  //       [questionId]: questionData
-  //     };
-  //   },{})
-  // }
 
   // handleValueUpdate = (questionId, questionData) => {
   //   this.setState((previousState) => {
@@ -87,6 +82,67 @@ class RandomAudioActivity extends Component {
         url="http://urlserveurs.free.fr/sound/misc/ooorgle.wav"
         playStatus={this.state.audioState}
       />
+      {
+        this.props.activity.questions.map((question, idx) => {
+            const questionId = idFromString(question.question);
+            if (question.type === 'text') {
+              return (
+                <div key={idx}>
+                  <ControlLabel>{question.question}</ControlLabel>
+                  <StyledFormControl
+                    type="text"
+                    value={get(this.state, `questions[${questionId}].response`) || ""}
+                    // onChange={
+                    //   (event) => this.handleValueUpdate(
+                    //     questionId,
+                    //     {
+                    //       indexInActivity: idx,
+                    //       type: "text",
+                    //       response: event.target.value
+                    //     }
+                    //   )
+                    // }
+                    />
+                </div>
+              );
+            } else if (question.type === 'select') {
+              return (
+                <div key={idx}>
+                  <ControlLabel>{question.question}</ControlLabel>
+                  <StyledFormControl
+                    componentClass="select"
+                    type="select"
+                    value={
+                      get(this.state, `questions[${questionId}].response`, '')
+                    }
+                    // onChange={
+                    //   (event) => this.handleValueUpdate(
+                    //     questionId,
+                    //     {
+                    //       indexInActivity: idx,
+                    //       type: "select",
+                    //       response: event.target.value
+                    //     }
+                    //   )
+                    // }
+                  >
+                    {question.options.map((option, idx) => {
+                      return (
+                        <option
+                          key={idx}
+                          value={idFromString(option)}
+                        >
+                          {option}
+                        </option>
+                      );
+                    })}
+                  </StyledFormControl>
+                </div>
+              );
+            }
+            return <p key={idx}>{JSON.stringify(question)}</p>;
+          })
+        }
 
     </div>;
   }